requirements.txt
Flask
Jinja2
Flask-Assets
flask-bcrypt
webassets
jsmin
cssmin
django-htmlmin